+++ zanata-war/src/main/java/org/zanata/webtrans/client/view/EditorOptionsDisplay.java
@@ -25,7 +25,7 @@ public interface EditorOptionsDisplay extends WidgetDisplay {
 
         void onEditorButtonsOptionChanged(Boolean editorButtons);
 
-        void onUseCodeMirrorOptionChanged(Boolean useCodeMirrorChkValue);
+        void onUseCodeMirrorOptionChanged(Boolean useCodeMirrorChkValue, boolean force);
 
         void onShowSaveApprovedWarningChanged(Boolean showSaveApprovedWarning);